Move documentation component in separate directory for reusability.
