+++ coral/src/app/components/documentation/utils/topic-documentation-helper.ts
