+++ coral/src/app/features/topics/details/documentation/components/NoDocumentationBanner.test.tsx
@@ -1,10 +1,15 @@
 import { NoDocumentationBanner } from "src/app/features/topics/details/documentation/components/NoDocumentationBanner";
 import { cleanup, render, screen } from "@testing-library/react";
+import userEvent from "@testing-library/user-event";
+
+const testAddDocumentation = jest.fn();
 
 describe("NoDocumentationBanner", () => {
+  const user = userEvent.setup();
+
   describe("shows all necessary elements", () => {
     beforeAll(() => {
-      render(<NoDocumentationBanner />);
+      render(<NoDocumentationBanner addDocumentation={testAddDocumentation} />);
     });
 
     afterAll(cleanup);
@@ -26,6 +31,24 @@ describe("NoDocumentationBanner", () => {
     it("shows button to add a documentation", () => {
       const button = screen.getByRole("button", { name: "Add documentation" });
       expect(button).toBeVisible();
+    });
+  });
+
+  describe("triggers a addDocumentation event", () => {
+    beforeEach(() => {
+      render(<NoDocumentationBanner addDocumentation={testAddDocumentation} />);
+    });
+
+    afterEach(() => {
+      cleanup();
+      jest.clearAllMocks();
+    });
+
+    it("adds documentation when user clicks button", async () => {
+      const button = screen.getByRole("button", { name: "Add documentation" });
+      await user.click(button);
+
+      expect(testAddDocumentation).toHaveBeenCalled();
     });
   });
 });
